refactor-#109: move generate schema to helpers
