+++ src/graphql/helpers/generate-gql-schema.ts
@@ -1,0 +1,18 @@
+import { NestFactory } from '@nestjs/core';
+import {
+  GraphQLSchemaBuilderModule,
+  GraphQLSchemaFactory,
+} from '@nestjs/graphql';
+import { printSchema } from 'graphql';
+
+import { MicroservicesModules } from '../../microservices';
+
+async function generateGqlSchema() {
+  const app = await NestFactory.create(GraphQLSchemaBuilderModule);
+  await app.init();
+
+  const gqlSchemaFactory = app.get(GraphQLSchemaFactory);
+  const schema = await gqlSchemaFactory.create(MicroservicesModules);
+
+  console.log(printSchema(schema));
+}
