+++ src/graphql/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './generate-gql-schema';
